requirements.txt
grpcio
chirpstack-api
tkinter
